+++ tests/testthat/test-writelas.R
@@ -5,7 +5,7 @@ las = readlasdata(lazfile)
 header = readlasheader(lazfile)
 write_path = file.path(tempdir(), "temp.laz")
 
-test_that("write.las write a correct file",{
+test_that("write.las writes a correct file",{
   write.las(write_path, header, las)
   wlas = readlasdata(write_path)
   wheader = readlasheader(write_path)
@@ -14,7 +14,7 @@ test_that("write.las write a correct file",{
   expect_equal(header$`Point Data Format ID`, wheader$`Point Data Format ID`)
 })
 
-test_that("write.las do not write time if point format enforced to 0",{
+test_that("write.las does not write time if point format enforced to 0",{
   new_header = header
   new_header$`Point Data Format ID` = 0
 
@@ -53,10 +53,7 @@ test_that("write.las writes Extra Bytes correctly",{
 
   expect_true(all(EB_header_check))
 
-  data1 = wlas[,.(X, Y, Z, Amplitude, `Pulse width`)]
-  data2 = las[,.(X, Y, Z, Amplitude, `Pulse width`)]
-
-  expect_equal(data1, data2)
+  expect_equal(las, wlas)
 })
 
 
@@ -68,7 +65,7 @@ test_that("write.las skip extra bytes if empty VLR", {
   wheader = readlasheader(write_path)
 
   expect_true(!any(c("Amplitude", "Pulse width") %in% names(wlas)))
-  expect_equal(wlas, las[, !"Amplitude", with=FALSE][, !"Pulse width", with=FALSE])
+  expect_equal(wlas, las[, -c(14:15)])
   expect_equal(length(wheader$`Variable Length Records`), 0)
 })
 
@@ -82,7 +79,7 @@ test_that("write.las skiped selectively extra byte if missing VLR",{
   wheader <- readlasheader(write_path)
 
   expect_true(!"Amplitude" %in% names(wlas))
-  expect_equal(wlas, las[, -"Amplitude", with = FALSE])
+  expect_equal(wlas, las[, -c(14)])
 })
 
 test_that("Modifying Extra Byte format works",{
@@ -99,12 +96,3 @@ test_that("Modifying Extra Byte format works",{
   expect_true(all(wlas$Amplitude == floor(las$Amplitude+.5)))
   expect_true(wheader$`Variable Length Records`$Extra_Bytes$`Extra Bytes Description`$Amplitude$data_type==1)
 })
-
-
-
-
-
-
-
-
-